Refactor OAuth code out of toolset_mcp.py into dedicated modules
Extract ~200 lines of OAuth infrastructure from toolset_mcp.py into
well-structured classes:

- oauth_config.py: MCPOAuthConfig, OAuthEndpoints, exceptions,
  OAuthExchangeManager (encapsulates pending exchange state)
- oauth_server_callbacks.py: server endpoint helpers (get_toolset_oauth_config,
  process_oauth_callback)
- oauth_tools_cache.py: OAuthToolsCache class (per-user tool cache,
  load_authenticated_oauth_tools, has_oauth_mcp_toolsets)

Also:
- Accept HTTP 201 in token exchange (Supabase returns 201 Created)
- Support client_secret in token exchange for confidential OAuth clients
- Move get_robusta_global_config_value to Config static method
- Remove signing_key getter indirection from OAuthTokenManager
